+++ src/main/java/prography/spring/pingpong/domain/room/model/dto/RoomDetailResponseDto.java
@@ -12,9 +12,9 @@ import prography.spring.pingpong.model.entity.RoomType;
 @Getter
 @Builder
 public class RoomDetailResponseDto {
-    private Long id;
+    private Integer id;
     private String title;
-    private Long hostId;
+    private Integer hostId;
     private RoomType roomType;
     private RoomStatus status;
