+++ src/main/java/prography/spring/pingpong/domain/userroom/model/dto/RoomExitRequestDto.java
@@ -1,4 +1,4 @@
 package prography.spring.pingpong.domain.userroom.model.dto;
 
-public record RoomExitRequestDto(Long userId) {
+public record RoomExitRequestDto(int userId) {
 }
